phase4-redshift/etl-functions/requirements.txt
boto3>=1.34.0
psycopg2-binary>=2.9.9
python-dateutil>=2.9.0
